+++ protected/humhub/docs/CHANGELOG.md
@@ -13,6 +13,8 @@ HumHub Change Log
 - Enh: Added `humhub\libs\RestrictedCallException`
 - Chng: Switched from [bootstrap-tour](https://github.com/sorich87/bootstrap-tour) to [bootstrap-tourist](https://github.com/IGreatlyDislikeJavascript/bootstrap-tourist) due to incompatibility to bootstrap v3.4.1
 - Enh: Added `humhub.modules.tour` module for handling tour logic
+- Fix: Added PHP 7.3 compatibility workaround for ZF2 ArrayObject
+
 
 1.3.12  (March 26, 2019)
 ---------------------------