+++ protected/humhub/docs/CHANGELOG.md
@@ -14,6 +14,7 @@ HumHub Change Log
 - Chng: Switched from [bootstrap-tour](https://github.com/sorich87/bootstrap-tour) to [bootstrap-tourist](https://github.com/IGreatlyDislikeJavascript/bootstrap-tourist) due to incompatibility to bootstrap v3.4.1
 - Enh: Added `humhub.modules.tour` module for handling tour logic
 - Fix: Added PHP 7.3 compatibility workaround for ZF2 ArrayObject
+- Fix: #3513 Profile schema cache not refreshed after adding new profile field
 
 
 1.3.12  (March 26, 2019)